Build fixes for the Percona 8.4.10 merge (inherited, macOS/portability)
None of these are merge regressions. Four separate breakages that block the
build on macOS/arm64; all verified against the parents before changing anything.

components/audit_log_filter/audit_udf.cc, .../log_writer/file_handle.cc and the
two sprintf sites in sql/mysqld.cc are byte-identical to Percona (git diff
against Percona-Server-8.4.10-10 is empty for the first two; for mysqld.cc,
which we did resolve conflicts in, the offending lines appear verbatim in
Percona at :11836/:11846 and origin/main contains zero occurrences of the
format string).

  Format-string mismatches: int64_t/uint64_t are 'long' on LP64 Linux but
  'long long' on macOS/arm64, so %ld/%lu trip -Werror,-Wformat. Percona's CI is
  Linux-only, so these never fire for them. Fixed portably with a cast to the
  widest type rather than a platform #ifdef.

  file_handle.cc: flush_on_write is only consumed inside #ifdef O_DIRECT.
  macOS has no O_DIRECT, so the #else branch compiles and the parameter is
  genuinely unused -> -Werror,-Wunused-parameter. Marked [[maybe_unused]],
  matching the style already used elsewhere in that component.

  unittest/gunit/CMakeLists.txt: a *configure*-time error, not a compile error.
  sqlgunitlib is linked with the keyword signature on one line and Percona's
  added line used the plain signature; CMake forbids mixing on one target. The
  file auto-merged cleanly -- the clash was semantic. The 8.4.8 merge hit this
  identical failure.

Worth reporting the format-string ones upstream to Percona; if they take them,
our divergence in that component goes to zero at the next merge.

+++ components/audit_log_filter/audit_udf.cc
@@ -1060,8 +1060,12 @@ char *AuditUdf::audit_log_read_bookmark_udf(AuditUdf *udf [[maybe_unused]],
                                             unsigned char *error) noexcept {
   auto bookmark = SysVars::get_log_bookmark();
 
-  std::snprintf(result, MYSQL_ERRMSG_SIZE, R"({"timestamp": "%s", "id": %lu})",
-                bookmark.timestamp.c_str(), bookmark.id);
+  // bookmark.id is uint64_t, which is 'unsigned long' on LP64 Linux but
+  // 'unsigned long long' on macOS/arm64, so a bare %lu fails -Werror=format
+  // there. Cast to the widest unsigned type so one format string works on both.
+  std::snprintf(result, MYSQL_ERRMSG_SIZE, R"({"timestamp": "%s", "id": %llu})",
+                bookmark.timestamp.c_str(),
+                static_cast<unsigned long long>(bookmark.id));
 
   *length = std::strlen(result);
   *is_null = 0;
